requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
decouple==0.0.7
defusedxml==0.7.1
distlib==0.3.5
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-googledrive-storage==1.6.0
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.1.0
filelock==3.8.0
Flask==2.1.2
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
google-api-core==2.10.1
google-api-python-client==2.61.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
heroku==0.1.4
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
neat-python==0.92
oauthlib==3.2.0
openpyxl==3.0.10
outcome==1.1.0
Pillow==9.2.0
platformdirs==2.5.2
protobuf==4.21.6
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.2
PyJWT==2.4.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
stripe==4.1.0
trio==0.20.0
trio-websocket==0.9.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.16.3
Werkzeug==2.1.2
whitenoise==6.2.0
wsproto==1.1.0
